[Bronze V] Title: Multiply, Time: 168 ms, Memory: 17580 KB, Score: 40 point -BaekjoonHub

+++ 백준/Bronze/22193. Multiply/Multiply.java
@@ -6,8 +6,8 @@ public class Main {
         Scanner sc = new Scanner(System.in);
         int n = sc.nextInt();
         int m = sc.nextInt();
-        int a = sc.nextInt();
-        int b = sc.nextInt();
+        long a = sc.nextLong();
+        long b = sc.nextLong();
         
         System.out.print(a*b);
     }
